feat: added new svg for sidebar filled and outlined
