+++ icons/jsx/index.jsx
@@ -1694,6 +1694,8 @@ export { default as RestoreFromTrash } from "./RestoreFromTrash";
 export { default as RestorePage } from "./RestorePage";
 export { default as Reviews } from "./Reviews";
 export { default as RiceBowl } from "./RiceBowl";
+export { default as RightSidebarFilled } from "./RightSidebarFilled";
+export { default as RightSidebarOutlined } from "./RightSidebarOutlined";
 export { default as RingVolume } from "./RingVolume";
 export { default as Rocket } from "./Rocket";
 export { default as RocketLaunch } from "./RocketLaunch";
@@ -2252,7 +2254,6 @@ export { default as Wc } from "./Wc";
 export { default as Web } from "./Web";
 export { default as WebAsset } from "./WebAsset";
 export { default as WebAssetOff } from "./WebAssetOff";
-export { default as WebBlack } from "./WebBlack";
 export { default as WebStories } from "./WebStories";
 export { default as Webhook } from "./Webhook";
 export { default as Wechat } from "./Wechat";
